+++ .mocharc.js
@@ -1,6 +1,6 @@
 module.exports = {
   'require': ['source-map-support/register', 'ts-node/register'],
   'recursive': true,
-  'spec': 'test/**/*.test.ts',
+  'spec': ['test/common/**/*.test.ts', 'test/node/**/*.test.ts'],
   'watch-extensions': ['ts'],
 };
