+++ karma.conf.js
@@ -15,7 +15,7 @@ module.exports = function(config) {
     // list of files / patterns to load in the browser
     files: [
       { pattern: 'src/**/*.ts' },
-      { pattern: 'test/**/*.ts' },
+      { pattern: 'test/common/**/*.ts' },
     ],
